RFI protection

+++ private/plugins/mediagallery/sql/sql_defaults.php
@@ -26,7 +26,10 @@
 // |                                                                           |
 // +---------------------------------------------------------------------------+
 //
-
+if (!defined ('GVERSION')) {
+{
+    die ('This file can not be used on its own.');
+}
 // Must set $ftp_path and $tmp_path in main routine before including this file.
 
 $_SQL_DEF[]="DELETE FROM {$_TABLES['mg_config']}";
